+++ Data/Scripts/OreDetectorReforged/Main.cs
@@ -1,7 +1,4 @@
-﻿using Sandbox.Common.ObjectBuilders;
-using Sandbox.Game.Entities;
-using Sandbox.Game.Entities.Cube;
-using Sandbox.Game.Gui;
+﻿using Sandbox.Game.Entities;
 using Sandbox.ModAPI;
 using System;
 using VRage.Game.Components;